+++ package/CHANGELOG.md
@@ -3,7 +3,8 @@
 - New modules: Force Enrage, Resize Enemies, Edges (World)
 - Made viewmodel tweaks work with arms
 - Added 12h/24h option for ClickGUI top bar clock
-- Fixed tooltip positioning that made buttons unclickable on (it seems) ultrawide monitor
+- Fixed hasToggling constructor param for Modules not working
+- Fixed tooltip positioning that made buttons unclickable in certain cases
 - Fixed very inconsistent icon colorizing
 - Fixed dropdown texts being empty